requirements.txt
langchain
langchain-groq
groq
langgraph

fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
requests
openai
google-generativeai
numpy
pandas
python-dotenv
reportlab
feedparser


typer
rich
prompt-toolkit
